refactor some dead code, add enemies folder
